modifications

+++ Timers/Timer1_reg.h
@@ -30,19 +30,19 @@
 
 #define TCNT1H				*((volatile u8*)0x4D)
 #define TCNT1L				*((volatile u8*)0x4C)
-#define TCNT1L_H				*((volatile u8*)0x4C)
+#define TCNT1L_H				*((volatile u16*)0x4C)
 
 #define OCR1AH 			*((volatile u8*)0x4B)
 #define OCR1AL 			*((volatile u8*)0x4A)
-#define OCR1AL_H 			*((volatile u8*)0x4A)
+#define OCR1AL_H 			*((volatile u16*)0x4A)
 
 #define OCR1BH 			*((volatile u8*)0x49)
 #define OCR1BL 			*((volatile u8*)0x48)
-#define OCR1BL_H 			*((volatile u8*)0x48)
+#define OCR1BL_H 			*((volatile u16*)0x48)
 
 #define ICR1H			*((volatile u8*)0x47)
 #define ICR1L			*((volatile u8*)0x46)
-#define ICR1L_H			*((volatile u8*)0x46)
+#define ICR1L_H			*((volatile u16*)0x46)
 
 #define TIMSK			*((volatile u8*)0x59)
 #define TIMSK_TOIE1					2
